refactor:  broke down the complex monolithic converter into manageable, focused modules
- preserving all existing functionality and maintaining the original public API surface.
- Introduced XamlContentAnalyzer for validating and analyzing XAML elements.
- Added XamlElementProcessor to handle processing of XAML elements into markdown.
- Enhanced XamlToMarkdownConverter to utilize new content analysis and element processing capabilities.
- Updated XamlTextFormatter with methods for decoding entities and formatting text.
- Refactored XamlListProcessor to integrate with the new element processor for improved list handling.
